// src/component/ecommerce/product/edit/schema.ts
import { z } from 'zod';
import { messages } from '@/config/messages';
import { fileSchema } from '@/utils/validators/common-rules';

const colorSchema = z.object({
  name: z.string(),
  code: z.string(),
  available: z.boolean(),
});

const sizeSchema = z.object({
  size: z.string(),
  available: z.boolean(),
});

export const productFormSchema = z.object({
  name: z.string().min(1, { message: messages.productNameIsRequired }),
  price: z.coerce.number().min(1, { message: messages.priceIsRequired }),
  mrp: z.coerce.number().min(1, { message: 'MRP is required' }),
  brand: z.string().min(1, { message: 'Brand is required' }),
  category: z.string().min(1, { message: 'Category is required' }),
  desc: z.string().min(1, { message: 'Description is required' }),
  images: z.array(fileSchema).optional(),
  stock: z.string().min(1, { message: 'Stock is required' }),
  tags: z.array(z.string()).optional(),
  colors: z.array(colorSchema).optional(),
  sizes: z.array(sizeSchema).optional(),
  instaId: z.string().min(1, { message: 'Instagram Post Link is required' }),
  height: z.string().min(1, { message: 'Height is required' }),
  weight: z.string().min(1, { message: 'Weight is required' }),
  length: z.string().min(1, { message: 'Length is required' }),
  breadth: z.string().min(1, { message: 'Breadth is required' }),
});

export type CreateProductInput = z.infer<typeof productFormSchema>;